+++ ElvUI_GennUI/modules/tooltipicon.lua
@@ -4,7 +4,7 @@ local GNUI = E:GetModule("GennUI");
 local GetAddOnMetadata = C_AddOns and C_AddOns.GetAddOnMetadata or GetAddOnMetadata
 
 --[[ Credit: brykrys, Alason, Freddy, Amavana, Resike, Merathilis ]]--
-local VERSION = 1.86
+local VERSION = 1.90
 local VERSIONINFO = "X-Release"
 
 local NEWTOOLTIPS = (TooltipDataProcessor and TooltipDataProcessor.AddTooltipPostCall) and true or false
@@ -76,7 +76,7 @@ local function DefaultSavedVariables()
 			size = 39,
 			alpha = 1,
 		},
-		inside = {size = 32},
+		inside = {size = 36},
 		background = {
 			alpha = .9,
 			tintr = .4,
@@ -608,6 +608,35 @@ local function HookMerchantItem(frame, index)
 	DisplayIconDispatch(data, tpath)
 end
 
+-- Handler for Macro datatable from TooltipDataProcessor
+-- Parse datatable for item or spell ID and find related icon
+local function HookMacro(frame, datatable)
+	local data = IconDataTable[frame]
+	if not data or data.disable or data.shown then
+		return
+	end
+	local info = datatable.lines[1] -- Assume lines always exists at this point, and that the info we need is in lines[1]
+	local tooltipType = info.tooltipType
+	local tooltipID = info.tooltipID
+	if tooltipType == Enum.TooltipDataType.Item then
+		if not options.item then
+			return
+		end
+		local icon = GetItemIcon(tooltipID)
+		if icon then
+			DisplayIconDispatch(data, icon)
+		end
+	elseif tooltipType == Enum.TooltipDataType.Spell then
+		if not options.spell then
+			return
+		end
+		local _, _, icon = GetSpellInfo(tooltipID)
+		if icon then
+			DisplayIconDispatch(data, icon)
+		end
+	end
+end
+
 -- Hook for frame:SetHyperlink
 local function HookHyperlink(frame, link)
 	local data = IconDataTable[frame]
@@ -1116,6 +1145,7 @@ local function OnEvent(frame) -- only event is VARIABLES_LOADED
 		TooltipDataProcessor.AddTooltipPostCall(Enum.TooltipDataType.Item, HookItem)
 		TooltipDataProcessor.AddTooltipPostCall(Enum.TooltipDataType.Spell, HookSpell)
 		TooltipDataProcessor.AddTooltipPostCall(Enum.TooltipDataType.EquipmentSet, HookEquipmentSet)
+		TooltipDataProcessor.AddTooltipPostCall(Enum.TooltipDataType.Macro, HookMacro)
 	end
 
 	-- slash commands